+++ library/smoothstreaming/src/test/java/com/google/android/exoplayer2/source/smoothstreaming/SsMediaSourceTest.java
@@ -20,9 +20,7 @@ import static com.google.common.truth.Truth.assertThat;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.android.exoplayer2.MediaItem;
-import com.google.android.exoplayer2.offline.StreamKey;
 import com.google.android.exoplayer2.upstream.FileDataSource;
-import java.util.Collections;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -64,46 +62,5 @@ public class SsMediaSourceTest {
     assertThat(ssMediaItem.localConfiguration.uri)
         .isEqualTo(castNonNull(mediaItem.localConfiguration).uri);
     assertThat(ssMediaItem.localConfiguration.tag).isEqualTo(mediaItemTag);
-  }
-
-  // Tests backwards compatibility
-  @SuppressWarnings("deprecation")
-  @Test
-  public void factorySetStreamKeys_emptyMediaItemStreamKeys_setsMediaItemStreamKeys() {
-    MediaItem mediaItem = MediaItem.fromUri("http://www.google.com");
-    StreamKey streamKey = new StreamKey(/* groupIndex= */ 0, /* trackIndex= */ 1);
-    SsMediaSource.Factory factory =
-        new SsMediaSource.Factory(new FileDataSource.Factory())
-            .setStreamKeys(Collections.singletonList(streamKey));
-
-    MediaItem ssMediaItem = factory.createMediaSource(mediaItem).getMediaItem();
-
-    assertThat(ssMediaItem.localConfiguration).isNotNull();
-    assertThat(ssMediaItem.localConfiguration.uri)
-        .isEqualTo(castNonNull(mediaItem.localConfiguration).uri);
-    assertThat(ssMediaItem.localConfiguration.streamKeys).containsExactly(streamKey);
-  }
-
-  // Tests backwards compatibility
-  @SuppressWarnings("deprecation")
-  @Test
-  public void factorySetStreamKeys_withMediaItemStreamKeys_doesNotOverrideMediaItemStreamKeys() {
-    StreamKey mediaItemStreamKey = new StreamKey(/* groupIndex= */ 0, /* trackIndex= */ 1);
-    MediaItem mediaItem =
-        new MediaItem.Builder()
-            .setUri("http://www.google.com")
-            .setStreamKeys(Collections.singletonList(mediaItemStreamKey))
-            .build();
-    SsMediaSource.Factory factory =
-        new SsMediaSource.Factory(new FileDataSource.Factory())
-            .setStreamKeys(
-                Collections.singletonList(new StreamKey(/* groupIndex= */ 1, /* trackIndex= */ 0)));
-
-    MediaItem ssMediaItem = factory.createMediaSource(mediaItem).getMediaItem();
-
-    assertThat(ssMediaItem.localConfiguration).isNotNull();
-    assertThat(ssMediaItem.localConfiguration.uri)
-        .isEqualTo(castNonNull(mediaItem.localConfiguration).uri);
-    assertThat(ssMediaItem.localConfiguration.streamKeys).containsExactly(mediaItemStreamKey);
   }
 }